requirements.txt
langchain == 0.2.11
langchain-openai == 0.1.17
streamlit == 1.36.0
python-dotenv == 1.0.1
PyPDF2 == 3.0.1